Started fixing a security issue
If the user doesn't have a https connection to the server someone can re send the url + cookies to the server to re-set older data a user has send send to the server if someone has captured server traffic. I found the encrytion already includes salt so i only need to check if the url has already been send earlier to check if this url is valid.